requirements.txt
python-pptx==0.6.21
Pillow==10.0.0
transformers==4.30.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.1
pytesseract==0.3.10
opencv-python-headless==4.8.0.76
numpy>=1.26.0
python-dotenv==1.0.0
setuptools>=68.0.0
wheel>=0.40.0
tokenizers==0.13.3 --no-build-isolation 